requirements.txt
# --------- base-tool --------- #
sphinx==8.1.3
sphinx-rtd-theme==3.0.2
pytest==8.3.5
pytest-cov==6.0.0
pyrootutils==1.0.4
pre-commit==4.1.0
twine==6.1.0
wheel>=0.45.1
coverage==7.6.12
furo==2024.8.6
myst-parser==4.0.1
pytest-mock==3.14.0
pendulum
python-dotenv==1.0.1
cookiecutter
setuptools
click

# --------- date --------- #
openpyxl==3.1.5
attrs>=25.1.0,<26

# --------- path --------- #
pathlib


# --------- cli --------- #
# No entries


# --------- ai-tools --------- #
mem0ai
openai==1.65.4
instructor


# --------- deploy --------- #
# ansible is commented out

# --------- data --------- #
numpy==1.26.4
pandas
pypdf==4.3.1
pyarrow==17.0.0
python-docx==1.1.2
python-magic==0.4.27
python-pptx==1.0.2
docx2txt==0.8
markdown==3.7


# --------- files --------- #
pyyaml
toml

# --------- mcp server --------- #
mcp
fastapi
fastmcp

# --------- db --------- #
chromadb

# --------- distributed --------- #
dora-rs-cli






